first-training/requirements.txt
blinker==1.9.0 ; python_version >= "3.13" and python_version < "4.0"
click==8.2.0 ; python_version >= "3.13" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.13" and python_version < "4.0" and platform_system == "Windows"
flask==3.1.1 ; python_version >= "3.13" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.13" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.13" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.13" and python_version < "4.0"
mypy-extensions==1.1.0 ; python_version >= "3.13" and python_version < "4.0"
mypy==1.15.0 ; python_version >= "3.13" and python_version < "4.0"
python-dotenv==1.1.0 ; python_version >= "3.13" and python_version < "4.0"
typing-extensions==4.13.2 ; python_version >= "3.13" and python_version < "4.0"
werkzeug==3.1.3 ; python_version >= "3.13" and python_version < "4.0"
